+++ src/test/kotlin/devtests/StringCases.kt
@@ -1,0 +1,12 @@
+package devtests
+
+import klib.extensions.toSnakeCase
+
+fun main() {
+    val normal = "This is a Test"
+
+    val snakeCased = normal.toSnakeCase()
+
+    println(normal)
+    println(snakeCased)
+}